requirements.txt
smartapi-python
websocket-client
pandas
numpy
scikit-learn
ta
python-dotenv
schedule
requests
pyotp
logzero
